+++ app/src/androidTest/java/de/mw136/tonuino/ExampleInstrumentedTest.kt
@@ -1,8 +1,8 @@
 package de.mw136.tonuino
 
 import android.os.Parcel
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import de.mw136.tonuino.nfc.TagData
 
 import org.junit.Test