+++ app/src/main/java/de/mw136/tonuino/nfc/NfcIntentActivity.kt
@@ -9,7 +9,7 @@ import android.nfc.Tag
 import android.nfc.tech.NfcA
 import android.nfc.tech.NfcB
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import android.widget.Toast
 import de.mw136.tonuino.R